pyproject.toml
[project]
name = "next-rush"
version = "0.1.0"
description = "Next RUSH+ Sonolus engine"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "sonolus-py~=0.12.6",
]

[dependency-groups]
dev = [
    "ruff",
    "pyright",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "W", "I", "N", "D", "UP", "YTT", "B", "A", "COM", "C4", "DTZ", "PIE", "PT", "Q", "SLOT", "SIM", "PTH", "PL", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E402", "E741", "D1", "COM812", "PLW2901", "PLW3201", "PLR6301", "PLC0415", "PLR2004", "PLR09", "SIM108", "FURB113", "A005", "B903", "FURB118", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.sources]
sonolus-py = { git = "https://github.com/qwewqa/sonolus.py.git", rev = "sonolus-1.1.0" }
